moving presentation layer files (viewModels, screens etc.) from the application module to the presentation module
