+++ presentation/src/main/java/com/fredprojects/helloworld/presentation/features/jumps/vm/JumpingRopeVM.kt
